fixed some critical physics bugs: - half-extents now computed in world space for box collider - physics components store a ref scale component (because the interpolated transform matrix from reactphysics3d only contains translation and rotation) - OBJModels now compute the object centers instead of the centroid (used for physics and normalizing the vertex buffer)

+++ src/Application.cpp
@@ -52,7 +52,7 @@ void Application::run()
 
 			float physicsTime = m_physics.InterpFactor(static_cast<float>(realFrameTime));
 
-			m_sceneManager.Update( physicsTime, realFrameTime );
+			m_sceneManager.Update( physicsTime, static_cast<float>(realFrameTime) );
 
 			m_vulkanGraphicsContext->Render(  m_sceneManager.GetSceneView() );
 		}
